[Transcription] Regenerate published samples for preview release
Regenerate samples/ outputs from samples-dev via 'dev-tool samples publish -f'. The preview version (1.1.0-beta.1) outputs to samples/v1-beta, replacing the stale GA samples/v1. Includes the new Enhanced Mode locale-hint example.
