+++ sdk/transcription/ai-speech-transcription/samples/v1-beta/javascript/multilingualTranscription.js
