+++ sdk/transcription/ai-speech-transcription/samples/v1-beta/javascript/package.json
@@ -1,14 +1,10 @@
 {
-  "name": "@azure-samples/ai-speech-transcription-ts",
+  "name": "@azure-samples/ai-speech-transcription-js-beta",
   "private": true,
   "version": "1.0.0",
-  "description": "Azure AI Speech Transcription client library samples for TypeScript",
+  "description": "Azure AI Speech Transcription client library samples for JavaScript (Beta)",
   "engines": {
     "node": ">=20.0.0"
-  },
-  "scripts": {
-    "build": "tsc",
-    "prebuild": "rimraf dist/"
   },
   "repository": {
     "type": "git",
@@ -30,14 +26,11 @@
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/transcription/ai-speech-transcription",
   "dependencies": {
-    "@azure/ai-speech-transcription": "latest",
+    "@azure/ai-speech-transcription": "beta",
     "dotenv": "latest",
     "@azure/core-auth": "^1.9.0"
   },
   "devDependencies": {
-    "@types/node": "^20.0.0",
-    "cross-env": "latest",
-    "rimraf": "latest",
-    "typescript": "~6.0.2"
+    "cross-env": "latest"
   }
 }
