+++ sdk/transcription/ai-speech-transcription/samples/v1-beta/typescript/package.json
@@ -1,10 +1,14 @@
 {
-  "name": "@azure-samples/ai-speech-transcription-js",
+  "name": "@azure-samples/ai-speech-transcription-ts-beta",
   "private": true,
   "version": "1.0.0",
-  "description": "Azure AI Speech Transcription client library samples for JavaScript",
+  "description": "Azure AI Speech Transcription client library samples for TypeScript (Beta)",
   "engines": {
     "node": ">=20.0.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "prebuild": "rimraf dist/"
   },
   "repository": {
     "type": "git",
@@ -26,11 +30,14 @@
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/transcription/ai-speech-transcription",
   "dependencies": {
-    "@azure/ai-speech-transcription": "latest",
+    "@azure/ai-speech-transcription": "beta",
     "dotenv": "latest",
     "@azure/core-auth": "^1.9.0"
   },
   "devDependencies": {
-    "cross-env": "latest"
+    "@types/node": "^20.0.0",
+    "cross-env": "latest",
+    "rimraf": "latest",
+    "typescript": "~6.0.2"
   }
 }
