+++ sdk/transcription/ai-speech-transcription/samples/v1-beta/typescript/src/transcriptionWithLocale.ts
